build: go 1.27, latest dependencies
Bump go.mod to 1.27.0 and every direct dependency to its latest major:
docker client v28, go-redis v9 + bsm/redislock, etcd 3.7, grpc 1.83 with
regenerated protobuf bindings, urfave/cli v3, sentry-go 0.48, go-viper
mapstructure, pro-bing, distribution/reference and moby/go-archive.

The embedded etcd used by --embedded-storage now runs on
go.etcd.io/etcd/server/v3/embed instead of the etcd integration test
framework: store/etcdv3/embedded.Cluster starts a listener-less member
and hands out namespaced in-process clients, so *testing.T no longer
threads through calcium.New, NewStore, NewETCD and the plugin loader.
The docker v28 client rewrites Transport on the *http.Client it is
handed, so each engine gets its own copy of the shared client.

Mocks are regenerated with mockery v3 from .mockery.yml; golangci-lint
v2 (modernize, shadow, inline) is clean on linux and darwin.

+++ client/interceptor/types.go
@@ -9,7 +9,7 @@ import (
 
 // RetryOptions .
 type RetryOptions struct {
-	Max int
+	Max uint64
 }
 
 type retryStream struct {
